small optimization to batch.GetArg(EndTransaction)
