+++ roachpb/batch.go
@@ -60,10 +60,20 @@ func (ba *BatchRequest) IsRange() bool {
 	return (ba.flags() & isRange) != 0
 }
 
-// GetArg returns the first request of the given type, if possible.
+// GetArg returns a request of the given type if one is contained in the
+// Batch. The request returned is the first of its kind, with the exception
+// of EndTransaction, where it examines the very last request only.
 func (ba *BatchRequest) GetArg(method Method) (Request, bool) {
-	// TODO(tschottdorf): when looking for EndTransaction, just look at the
-	// last entry.
+	// when looking for EndTransaction, just look at the last entry.
+	if method == EndTransaction {
+		if length := len(ba.Requests); length > 0 {
+			if req := ba.Requests[length-1].GetInner(); req.Method() == EndTransaction {
+				return req, true
+			}
+		}
+		return nil, false
+	}
+
 	for _, arg := range ba.Requests {
 		if req := arg.GetInner(); req.Method() == method {
 			return req, true
